backend/requirements.txt
fastapi
uvicorn
python-dotenv
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-silero
livekit-api
openai
deepgram-sdk
elevenlabs
langchain
langchain-openai
langchain-chroma
chromadb
voyageai
mistralai
sqlalchemy
psycopg2-binary
boto3
pydantic-settings
supabase
pinecone-client
httpx
